Cart Quantity added

+++ server/apps/account/views.py
@@ -161,6 +161,10 @@ class CartView(generics.ListCreateAPIView):
         if not cartItem:
             return Response({"detail": "Cart item not found."}, status=status.HTTP_404_NOT_FOUND)
 
-        cartItem.delete()
+        if cartItem.quantity > 1:
+            cartItem.quantity -= 1
+            cartItem.save()
+        else:
+            cartItem.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
